% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAssignmentProps.R
\name{checkAssignmentProps}
\alias{checkAssignmentProps}
\title{Takes raw data from an experiment and tests the observed assignment splits are as expected}
\usage{
checkAssignmentProps(
  raw_data,
  sample_variable,
  control_value,
  expected_control_split
)
}
\description{
Takes raw data from an experiment and tests the observed assignment splits are as expected
}
\examples{
summarize_data()
}
\keyword{ab}
\keyword{summarise}
\keyword{test,}
