% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDoubleAssignment.R
\name{checkDoubleAssignment}
\alias{checkDoubleAssignment}
\title{Check if one user has multiple assignments across variations}
\usage{
checkDoubleAssignment(raw_data, id_variable, sample_variable, control_value)
}
\description{
Check if one user has multiple assignments across variations
}
\examples{
summarize_data()
}
\keyword{ab}
\keyword{summarise}
\keyword{test,}
