% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUniqueness.R
\name{checkUniqueness}
\alias{checkUniqueness}
\title{Check if one test subject has multiple observations within the data}
\usage{
checkUniqueness(raw_data, id_variable)
}
\description{
Check if one test subject has multiple observations within the data
}
\examples{
checkUniqueness(raw_test_data,user_id)
}
\keyword{ab}
\keyword{summarise}
\keyword{test,}
