% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDoubleAssignment.R
\name{removeDoubleAssignment}
\alias{removeDoubleAssignment}
\title{Check if one user has multiple variation assignments within data}
\usage{
removeDoubleAssignment(raw_data, id_variable, sample_variable, control_value)
}
\description{
Check if one user has multiple variation assignments within data
}
\examples{
summarize_data()
}
\keyword{ab}
\keyword{summarise}
\keyword{test,}
