% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeBayesianProp.R
\name{sizeBayesianProp}
\alias{sizeBayesianProp}
\title{takes a few inputs and returns a data.frame with various bayesian summary statistics for hypothetical deltas
and experiment run times}
\usage{
sizeBayesianProp(
  baseline,
  min_delta,
  max_delta,
  incremental_differences,
  n_per_day,
  sample_split,
  max_days,
  day_increments,
  min_effect_desized
)
}
\description{
takes a few inputs and returns a data.frame with various bayesian summary statistics for hypothetical deltas
and experiment run times
}
\examples{
summarize_data()
}
\keyword{ab}
\keyword{summarise}
\keyword{test,}
