% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseBayesianProp.R
\name{summariseBayesianProp}
\alias{summariseBayesianProp}
\title{Takes summary data from 'summariseMetrics' and generates Bayesiam Summary Statistics.  Output is a clean summary table!}
\usage{
summariseBayesianProp(data, probability_threshold)
}
\description{
Takes summary data from 'summariseMetrics' and generates Bayesiam Summary Statistics.  Output is a clean summary table!
}
\examples{
summariseBayesianProp(test_metric_agg,.01)
}
\keyword{ab}
\keyword{summarise}
\keyword{test,}
