% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseMetric.R
\name{summariseMetric}
\alias{summariseMetric}
\title{Takes a dataframe with a control and variant and summarizes proportions form user-inputted variables}
\usage{
summariseMetric(metric_data)
}
\description{
Takes a dataframe with a control and variant and summarizes proportions form user-inputted variables
}
\examples{
summarize_data()
}
\keyword{ab}
\keyword{summarise}
\keyword{test,}
