% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseMetricDaily.R
\name{summariseMetricDaily}
\alias{summariseMetricDaily}
\title{Takes a metric dataframe and summarizes it on a cumulative, daily level}
\usage{
summariseMetricDaily(metric_data)
}
\description{
Takes a metric dataframe and summarizes it on a cumulative, daily level
}
\examples{
summarize_data()
}
\keyword{ab}
\keyword{summarise}
\keyword{test,}
