% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeBayesianProp.R
\name{visualizeBayesianProp}
\alias{visualizeBayesianProp}
\title{This takes results from summariseMetric and visualizes the metrics within a Bayesian Framework}
\usage{
visualizeBayesianProp(data)
}
\description{
This takes results from summariseMetric and visualizes the metrics within a Bayesian Framework
}
\examples{
visualizeBayesianProp(test_metric_agg)
}
\keyword{ab}
\keyword{summarise}
\keyword{test,}
