% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeBayesianPropDaily.R
\name{visualizeBayesianPropDaily}
\alias{visualizeBayesianPropDaily}
\title{This takes results from summariseMetric and visualizes the metrics within a Bayesian Framework}
\usage{
visualizeBayesianPropDaily(daily_bayes)
}
\description{
This takes results from summariseMetric and visualizes the metrics within a Bayesian Framework
}
\examples{
summarize_data()
}
\keyword{ab}
\keyword{summarise}
\keyword{test,}
